Split of ev_io watcher for read and write
This commit splits up the single ev_io watcher that was previously used
for both read and write into two separate watchers. The reason for this
split is that a write event will always be triggered as soon as the
socket is writable which leads to busy-looping. Instead, the new
behavior is to start a separate ev_io watcher for writing as soon as a
message has been enqueued. The write watcher is disabled as soon as no
messages are left in the outgoing message queue of the client.